+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0] - 2023-09-22
+
+### Added
+
+- Default scopes for OAuth
+- New data from the API: Engine status
+
+### Changed
+
+- Adjusted the logic for downloading the header image
+
 ## [2.0.0] - 2023-07-31
 
 ### Added